[refactor] http 패키지 구조 수정
좀 더 알기 쉬운 패키지 구조로 수정하였습니다.
server 패키지와 http 패키지로 구조를 바꾸었습니다.
