+++ src/main/java/codesquad/http/HttpHandler.java
@@ -1,4 +1,4 @@
-package codesquad.httpHandler;
+package codesquad.http;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -10,9 +10,6 @@ import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import codesquad.httpRequest.HttpRequest;
-import codesquad.httpResponse.HttpResponse;
 
 public class HttpHandler {
